requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and NLP
transformers>=4.30.0
datasets>=2.12.0
sentence-transformers>=2.2.0
accelerate>=0.20.0
evaluate>=0.4.0

# Computer Vision
pillow>=9.5.0
opencv-python>=4.8.0
clip-by-openai>=1.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Training and Logging
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# NLP Processing
nltk>=3.8.0
spacy>=3.6.0
rouge-score>=0.1.2

# Knowledge Graphs
networkx>=3.1.0
rdflib>=6.3.0

# Utilities
pyyaml>=6.0
requests>=2.28.0
